pyproject.toml
[project]
name = "cthgithubactions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.38",
    "fastapi>=0.129.0",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
